refactor chrono_cache.rs to cli
