+++ benches/impl_path_string_for_evaluation_context.rs
@@ -10,7 +10,7 @@ use hyper::HeaderMap;
 use indexmap::IndexMap;
 use once_cell::sync::Lazy;
 use tailcall::blueprint::Server;
-use tailcall::chrono_cache::NativeChronoCache;
+use tailcall::cli::chrono_cache::NativeChronoCache;
 use tailcall::cli::{init_env, init_http, init_http2_only};
 use tailcall::http::RequestContext;
 use tailcall::lambda::{EvaluationContext, ResolverContextLike};
