+++ src/cli/mod.rs
@@ -1,3 +1,4 @@
+pub mod chrono_cache;
 mod command;
 mod error;
 mod fmt;
@@ -15,11 +16,10 @@ use crate::HttpIO;
 pub(crate) mod env;
 pub(crate) mod file;
 pub(crate) mod http;
+use chrono_cache::NativeChronoCache;
 pub use env::EnvNative;
 pub use file::NativeFileIO;
 pub use http::NativeHttp;
-
-use crate::chrono_cache::NativeChronoCache;
 
 // Provides access to env in native rust environment
 pub fn init_env() -> env::EnvNative {
