+++ src/lib.rs
@@ -4,7 +4,6 @@ mod app_context;
 pub mod async_graphql_hyper;
 pub mod blueprint;
 pub mod cache;
-pub mod chrono_cache;
 #[cfg(feature = "default")]
 pub mod cli;
 pub mod config;
